+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.2.1 (unreleased)
 
+- Added recursive scraping.
+
 ## v0.2.0
 
 - Added repsonsive grid.